diff parsing
Add coroutine based parsing of diff produced by OBS. This is a bit
special as it lists changed, new and removed files. The diff_to_dict
splits it to OrderedDict with name of a file (or special section) as a
key and list with lines of a path as a value.